+++ app/Http/Controllers/DanhMucController.php
@@ -21,7 +21,17 @@ class DanhMucController extends Controller
     public function getData()
     {
         $id_chuc_nang = 1;
-        
+        $login = Auth::guard('sanctum')->user();
+        $id_quyen = $login->$id_chuc_nang;
+        $check_quyen = ChiTietPhanQuyen::where('id_quyen', $id_quyen)
+            ->where('id_chuc_nang', $id_chuc_nang)
+            ->first();
+        if ($check_quyen) {
+            return response()->json([
+                'data' => false,
+                'message' => "bạn không có quyền thực hiện chức năng này!"
+            ]);
+        }
         $data = DanhMuc::get(); //Nghia la lay ra
 
         return response()->json([
@@ -33,7 +43,17 @@ class DanhMucController extends Controller
     {
 
         $id_chuc_nang = 2;
-
+        $login = Auth::guard('sanctum')->user();
+        $id_quyen = $login->$id_chuc_nang;
+        $check_quyen = ChiTietPhanQuyen::where('id_quyen', $id_quyen)
+            ->where('id_chuc_nang', $id_chuc_nang)
+            ->first();
+        if ($check_quyen) {
+            return response()->json([
+                'data' => false,
+                'message' => "bạn không có quyền thực hiện chức năng này!"
+            ]);
+        }
         DanhMuc::create([
             'ten_danh_muc' => $request->ten_danh_muc,
             'slug_danh_muc' => $request->slug_danh_muc,
@@ -51,7 +71,17 @@ class DanhMucController extends Controller
     {
 
         $id_chuc_nang = 4;
-
+        $login = Auth::guard('sanctum')->user();
+        $id_quyen = $login->$id_chuc_nang;
+        $check_quyen = ChiTietPhanQuyen::where('id_quyen', $id_quyen)
+            ->where('id_chuc_nang', $id_chuc_nang)
+            ->first();
+        if ($check_quyen) {
+            return response()->json([
+                'data' => false,
+                'message' => "bạn không có quyền thực hiện chức năng này!"
+            ]);
+        }
         //table danh mục tìm id = $request->id và sau đó xóa nó đi
         DanhMuc::find($request->id)->delete();
         return response()->json([
@@ -63,7 +93,17 @@ class DanhMucController extends Controller
     {
 
         $id_chuc_nang = 3;
-
+        $login = Auth::guard('sanctum')->user();
+        $id_quyen = $login->$id_chuc_nang;
+        $check_quyen = ChiTietPhanQuyen::where('id_quyen', $id_quyen)
+            ->where('id_chuc_nang', $id_chuc_nang)
+            ->first();
+        if ($check_quyen) {
+            return response()->json([
+                'data' => false,
+                'message' => "bạn không có quyền thực hiện chức năng này!"
+            ]);
+        }
         $slug = $request->slug_danh_muc;
         $check = DanhMuc::where('slug_danh_muc', $slug)->first();
         if ($check) {
@@ -82,7 +122,17 @@ class DanhMucController extends Controller
     {
 
         $id_chuc_nang = 5;
-
+        $login = Auth::guard('sanctum')->user();
+        $id_quyen = $login->$id_chuc_nang;
+        $check_quyen = ChiTietPhanQuyen::where('id_quyen', $id_quyen)
+            ->where('id_chuc_nang', $id_chuc_nang)
+            ->first();
+        if ($check_quyen) {
+            return response()->json([
+                'data' => false,
+                'message' => "bạn không có quyền thực hiện chức năng này!"
+            ]);
+        }
         DanhMuc::find($request->id)->update([
             'ten_danh_muc' => $request->ten_danh_muc,
             'slug_danh_muc' => $request->slug_danh_muc,
@@ -99,7 +149,17 @@ class DanhMucController extends Controller
     {
 
         $id_chuc_nang = 6;
-
+        $login = Auth::guard('sanctum')->user();
+        $id_quyen = $login->$id_chuc_nang;
+        $check_quyen = ChiTietPhanQuyen::where('id_quyen', $id_quyen)
+            ->where('id_chuc_nang', $id_chuc_nang)
+            ->first();
+        if ($check_quyen) {
+            return response()->json([
+                'data' => false,
+                'message' => "bạn không có quyền thực hiện chức năng này!"
+            ]);
+        }
         $danhMuc = DanhMuc::where('id', $request->id)->first();
 
         if ($danhMuc) {
