+++ app/Http/Controllers/SanPhamController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\ChiTietPhanQuyen;
 use App\Models\SanPham;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -32,7 +33,17 @@ class SanPhamController extends Controller
     {
         
         $id_chuc_nang = 12;
-        
+        $login = Auth::guard('sanctum')->user();
+        $id_quyen = $login->$id_chuc_nang;
+        $check_quyen = ChiTietPhanQuyen::where('id_quyen', $id_quyen)
+            ->where('id_chuc_nang', $id_chuc_nang)
+            ->first();
+        if ($check_quyen) {
+            return response()->json([
+                'data' => false,
+                'message' => "bạn không có quyền thực hiện chức năng này!"
+            ]);
+        }
         $tinh_trang = $request->tinh_trang == 1 ? 0 : 1;
         SanPham::find($request->id)->update([
             'tinh_trang'    =>  $tinh_trang
@@ -48,7 +59,17 @@ class SanPhamController extends Controller
     {
         
         $id_chuc_nang = 13;
-        
+        $login = Auth::guard('sanctum')->user();
+        $id_quyen = $login->$id_chuc_nang;
+        $check_quyen = ChiTietPhanQuyen::where('id_quyen', $id_quyen)
+            ->where('id_chuc_nang', $id_chuc_nang)
+            ->first();
+        if ($check_quyen) {
+            return response()->json([
+                'data' => false,
+                'message' => "bạn không có quyền thực hiện chức năng này!"
+            ]);
+        }
         $is_noi_bat = $request->is_noi_bat == 1 ? 0 : 1;
         SanPham::find($request->id)->update([
             'is_noi_bat'    =>  $is_noi_bat
@@ -63,7 +84,17 @@ class SanPhamController extends Controller
     public function update(Request $request){
         
         $id_chuc_nang = 11;
-        
+        $login = Auth::guard('sanctum')->user();
+        $id_quyen = $login->$id_chuc_nang;
+        $check_quyen = ChiTietPhanQuyen::where('id_quyen', $id_quyen)
+            ->where('id_chuc_nang', $id_chuc_nang)
+            ->first();
+        if ($check_quyen) {
+            return response()->json([
+                'data' => false,
+                'message' => "bạn không có quyền thực hiện chức năng này!"
+            ]);
+        }
         SanPham::find($request->id)->update([
             'ten_san_pham'  =>$request->ten_san_pham,
             'slug_san_pham'  =>$request->slug_san_pham,
@@ -86,7 +117,17 @@ class SanPhamController extends Controller
     {
         
         $id_chuc_nang = 14;
-        
+        $login = Auth::guard('sanctum')->user();
+        $id_quyen = $login->$id_chuc_nang;
+        $check_quyen = ChiTietPhanQuyen::where('id_quyen', $id_quyen)
+            ->where('id_chuc_nang', $id_chuc_nang)
+            ->first();
+        if ($check_quyen) {
+            return response()->json([
+                'data' => false,
+                'message' => "bạn không có quyền thực hiện chức năng này!"
+            ]);
+        }
         $is_flash_sale = $request->is_flash_sale == 1 ? 0 : 1;
         SanPham::find($request->id)->update([
             'is_flash_sale'    =>  $is_flash_sale
@@ -102,7 +143,17 @@ class SanPhamController extends Controller
     {
         
         $id_chuc_nang = 7;
-        
+        $login = Auth::guard('sanctum')->user();
+        $id_quyen = $login->$id_chuc_nang;
+        $check_quyen = ChiTietPhanQuyen::where('id_quyen', $id_quyen)
+            ->where('id_chuc_nang', $id_chuc_nang)
+            ->first();
+        if ($check_quyen) {
+            return response()->json([
+                'data' => false,
+                'message' => "bạn không có quyền thực hiện chức năng này!"
+            ]);
+        }
         $data = SanPham::get();
         return response()->json([
             'data' => $data
@@ -137,7 +188,17 @@ class SanPhamController extends Controller
     {
         
         $id_chuc_nang = 8;
-        
+        $login = Auth::guard('sanctum')->user();
+        $id_quyen = $login->$id_chuc_nang;
+        $check_quyen = ChiTietPhanQuyen::where('id_quyen', $id_quyen)
+            ->where('id_chuc_nang', $id_chuc_nang)
+            ->first();
+        if ($check_quyen) {
+            return response()->json([
+                'data' => false,
+                'message' => "bạn không có quyền thực hiện chức năng này!"
+            ]);
+        }
         SanPham::create([
             'ten_san_pham'      =>$request->ten_san_pham,
             'slug_san_pham'     =>$request->slug_san_pham,
@@ -161,7 +222,17 @@ class SanPhamController extends Controller
     {
         
         $id_chuc_nang = 9;
-        
+        $login = Auth::guard('sanctum')->user();
+        $id_quyen = $login->$id_chuc_nang;
+        $check_quyen = ChiTietPhanQuyen::where('id_quyen', $id_quyen)
+            ->where('id_chuc_nang', $id_chuc_nang)
+            ->first();
+        if ($check_quyen) {
+            return response()->json([
+                'data' => false,
+                'message' => "bạn không có quyền thực hiện chức năng này!"
+            ]);
+        }
         $slug = $request->slug_san_pham;
         $check = SanPham::where('slug_san_pham', $slug)->first();
         if($check){
@@ -181,7 +252,17 @@ class SanPhamController extends Controller
     {
         
         $id_chuc_nang = 10;
-        
+        $login = Auth::guard('sanctum')->user();
+        $id_quyen = $login->$id_chuc_nang;
+        $check_quyen = ChiTietPhanQuyen::where('id_quyen', $id_quyen)
+            ->where('id_chuc_nang', $id_chuc_nang)
+            ->first();
+        if ($check_quyen) {
+            return response()->json([
+                'data' => false,
+                'message' => "bạn không có quyền thực hiện chức năng này!"
+            ]);
+        }
         SanPham::find($request->id)->delete();
         return response()->json([
             'status' => true,
